Continue to optimize the test file
Add an auxiliary function 'format_request' for formatting the request message. This function ensures that the request message complies with the requirements of the protocol. In the client_thread function, the format_request function is used to format each request to ensure that the requests sent to the server are correct.